git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@20767 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ native/win32/peerident/com_l7tech_common_security_socket_Win32LocalTcpPeerIdentifier.cpp
@@ -201,7 +201,7 @@ report_error_auto(JNIEnv* env, const string prefix, DWORD err)
 bool 
 get_tcp_table(JNIEnv* env, vector<BYTE>& bytes)
 {
-	DWORD bytes_next_size = max(INITIAL_BUFFER_SIZE, bytes.capacity());
+	size_t bytes_next_size = max(INITIAL_BUFFER_SIZE, bytes.capacity());
 
 	for (;;) {
 		bytes.reserve(bytes_next_size);
